requirements.txt
gradio
pdfplumber
python-docx
docx2txt
python-dateutil
pandas